+++ yellowstone-grpc-geyser/src/grpc.rs
@@ -35,7 +35,7 @@ use {
         time::{sleep, Duration, Instant},
     },
     tonic::{
-        service::interceptor::interceptor,
+        service::interceptor,
         transport::{
             server::{Server, TcpIncoming},
             Identity, ServerTlsConfig,
@@ -393,12 +393,9 @@ impl GrpcService {
         Arc<Notify>,
     )> {
         // Bind service address
-        let incoming = TcpIncoming::new(
-            config.address,
-            true,                          // tcp_nodelay
-            Some(Duration::from_secs(20)), // tcp_keepalive
-        )
-        .map_err(|error| anyhow::anyhow!(error))?;
+        let incoming = TcpIncoming::bind(config.address)?
+            .with_nodelay(Some(true))
+            .with_keepalive(Some(Duration::from_secs(20)));
 
         // Snapshot channel
         let (snapshot_tx, snapshot_rx) = match config.snapshot_plugin_channel_capacity {
@@ -519,20 +516,22 @@ impl GrpcService {
         let shutdown_grpc = Arc::clone(&shutdown);
         tokio::spawn(async move {
             // gRPC Health check service
-            let (mut health_reporter, health_service) = health_reporter();
+            let (health_reporter, health_service) = health_reporter();
             health_reporter.set_serving::<GeyserServer<Self>>().await;
 
             server_builder
-                .layer(interceptor(move |request: Request<()>| {
-                    if let Some(x_token) = &config.x_token {
-                        match request.metadata().get("x-token") {
-                            Some(token) if x_token == token => Ok(request),
-                            _ => Err(Status::unauthenticated("No valid auth token")),
+                .layer(interceptor::InterceptorLayer::new(
+                    move |request: Request<()>| {
+                        if let Some(x_token) = &config.x_token {
+                            match request.metadata().get("x-token") {
+                                Some(token) if x_token == token => Ok(request),
+                                _ => Err(Status::unauthenticated("No valid auth token")),
+                            }
+                        } else {
+                            Ok(request)
                         }
-                    } else {
-                        Ok(request)
-                    }
-                }))
+                    },
+                ))
                 .add_service(health_service)
                 .add_service(service)
                 .serve_with_incoming_shutdown(incoming, shutdown_grpc.notified())