+++ yellowstone-grpc-proto/build.rs
@@ -1,19 +1,19 @@
 use {
     std::{env, fs, path::Path},
-    tonic_build::manual::{Builder, Method, Service},
+    tonic_prost_build::manual::{Builder, Method, Service},
 };
 
 fn main() -> anyhow::Result<()> {
     std::env::set_var("PROTOC", protobuf_src::protoc());
 
     // build protos
-    tonic_build::configure().compile_protos(&["proto/geyser.proto"], &["proto"])?;
+    tonic_prost_build::configure().compile_protos(&["proto/geyser.proto"], &["proto"])?;
 
     // build protos without tonic (wasm)
     let out_dir = env::var("OUT_DIR").expect("OUT_DIR not found");
     let out_dir_path = Path::new(&out_dir).join("no-tonic");
     fs::create_dir_all(&out_dir_path).expect("failed to create out no-tonic directory");
-    tonic_build::configure()
+    tonic_prost_build::configure()
         .build_client(false)
         .build_server(false)
         .out_dir(out_dir_path)
@@ -30,8 +30,7 @@ fn main() -> anyhow::Result<()> {
                 .input_type("crate::geyser::SubscribeRequest")
                 // .output_type("crate::geyser::SubscribeUpdate")
                 .output_type("crate::plugin::filter::message::FilteredUpdate")
-                .codec_path("tonic::codec::ProstCodec")
-                // .codec_path("crate::plugin::codec::SubscribeCodec")
+                .codec_path("tonic_prost::ProstCodec")
                 .client_streaming()
                 .server_streaming()
                 .build(),
@@ -42,7 +41,7 @@ fn main() -> anyhow::Result<()> {
                 .route_name("SubscribeReplayInfo")
                 .input_type("crate::geyser::SubscribeReplayInfoRequest")
                 .output_type("crate::geyser::SubscribeReplayInfoResponse")
-                .codec_path("tonic::codec::ProstCodec")
+                .codec_path("tonic_prost::ProstCodec")
                 .build(),
         )
         .method(
@@ -51,7 +50,7 @@ fn main() -> anyhow::Result<()> {
                 .route_name("Ping")
                 .input_type("crate::geyser::PingRequest")
                 .output_type("crate::geyser::PongResponse")
-                .codec_path("tonic::codec::ProstCodec")
+                .codec_path("tonic_prost::ProstCodec")
                 .build(),
         )
         .method(
@@ -60,7 +59,7 @@ fn main() -> anyhow::Result<()> {
                 .route_name("GetLatestBlockhash")
                 .input_type("crate::geyser::GetLatestBlockhashRequest")
                 .output_type("crate::geyser::GetLatestBlockhashResponse")
-                .codec_path("tonic::codec::ProstCodec")
+                .codec_path("tonic_prost::ProstCodec")
                 .build(),
         )
         .method(
@@ -69,7 +68,7 @@ fn main() -> anyhow::Result<()> {
                 .route_name("GetBlockHeight")
                 .input_type("crate::geyser::GetBlockHeightRequest")
                 .output_type("crate::geyser::GetBlockHeightResponse")
-                .codec_path("tonic::codec::ProstCodec")
+                .codec_path("tonic_prost::ProstCodec")
                 .build(),
         )
         .method(
@@ -78,7 +77,7 @@ fn main() -> anyhow::Result<()> {
                 .route_name("GetSlot")
                 .input_type("crate::geyser::GetSlotRequest")
                 .output_type("crate::geyser::GetSlotResponse")
-                .codec_path("tonic::codec::ProstCodec")
+                .codec_path("tonic_prost::ProstCodec")
                 .build(),
         )
         .method(
@@ -87,7 +86,7 @@ fn main() -> anyhow::Result<()> {
                 .route_name("IsBlockhashValid")
                 .input_type("crate::geyser::IsBlockhashValidRequest")
                 .output_type("crate::geyser::IsBlockhashValidResponse")
-                .codec_path("tonic::codec::ProstCodec")
+                .codec_path("tonic_prost::ProstCodec")
                 .build(),
         )
         .method(
@@ -96,7 +95,7 @@ fn main() -> anyhow::Result<()> {
                 .route_name("GetVersion")
                 .input_type("crate::geyser::GetVersionRequest")
                 .output_type("crate::geyser::GetVersionResponse")
-                .codec_path("tonic::codec::ProstCodec")
+                .codec_path("tonic_prost::ProstCodec")
                 .build(),
         )
         .build();