+++ CHANGELOG.md
@@ -16,7 +16,20 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
-## 2025-07-01
+
+## 2025-08-04
+
+- yellowstone-grpc-geyser-9.0.0
+- yellowstone-grpc-client-simple-9.0.0
+- yellowstone-grpc-client-9.0.0
+- yellowstone-grpc-geyser-9.0.0
+- yellowstone-grpc-proto-9.0.0
+
+### Breaking
+
+- Upgaded to tonic 0.14.0 which can break customer code if they rely on tonic to do something else.
+
+## 2025-08-01
 
 - yellowstone-grpc-geyser-8.1.0